[mlir][Linalg] Allow more control in drop unit dims
Extend the `ControlDropUnitDims` struct to allow users of the `linalg::dropUnitDims` function more control over the behavior of the function.

The extended struct allows users to specify functions to control how the operands are collapsed and how the result is expanded to the original shape.

One example (and the motivation for this change) where this additional control is useful is to allow collapsing of tensors with an encoding, as demonstrated by the new test.

This is a breaking change. The default behavior changes to abort the transformation if one of the operands cannot be collapsed. In the default implementation, this is the case for memref with non-identity layout and tensors with an encoding.

Signed-off-by: Lukas Sommer <lukas.sommer@amd.com>